requirements.txt
paho-mqtt
PyYAML
requests
